Parsing JSON starter code
Attempted to parse JSON objects to be displayed in a ListView using a LocationAdapter. Still not yet reading from newspaper_locations.json file (not showing up in Android project structure?). Next, I will be looking to read in json file directly. Then, I want to display style, details, lat/long for each item in a ScrollView and set the background based on the style in the ListView. Finally, I will make each item clickable and launch the GoogleMaps service to pinpoint the location based off the given latitude and longitude.